requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Pillow==10.0.1